Bookworm/test/requirements.txt
pymongo==3.12.0
pandas
redis
requests
pytest-html
jsonschema
scipy
sklearn
datetime
